Overhaul option parsing.
The callback mechanism is for advance parsing of arguments, not
necessarily for doing the work. For example you could pass --print
or --inject multiple times in the same command line invocation, which
seems strange to me.

The new model will print help if nothing is specified as well print out
metadata for all files listed as the default behaviour.

Fixes #3.