Bring back the "spherical" checkbox removed in 4ec34b2.
Since we removed the checkbox, we've learned that various tutorials and videos refer to checking "the spherical checkbox", or worse, checking "the first checkbox", so I'm bringing it back (albeit with a more descriptive label) to reduce confusion. As before, this checkbox doesn't actually do very much, since the app doesn't allow you to inject "non-spherical" metadata.